fix circular file calls
